requirements.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
dnspython>=2.2.1
python-whois>=0.7.3
colorama>=0.4.5
argparse>=1.4.0
tqdm>=4.64.0
urllib3>=1.26.10
cryptography>=38.0.1
pyOpenSSL>=22.0.0
shodan>=1.28.0
netaddr>=0.8.0
selenium>=4.4.3
webdriver-manager>=3.8.3
lxml>=4.9.1
aiohttp>=3.8.3
aiodns>=3.0.0
pycares>=4.2.2
jinja2>=3.1.2
pillow>=9.2.0
googlesearch-python>=1.1.0
python-masscan>=1.0.0
alive-progress>=2.4.1
click>=8.1.3
sslyze>=5.0.5
rich>=12.5.1
scrapy>=2.7.1
httpx>=0.23.0
uvloop>=0.17.0; sys_platform != 'win32'
asyncio>=3.4.3 